+++ lib/fetching_state.dart
@@ -3,22 +3,22 @@ library fetching_state;
 import 'package:flutter/material.dart';
 
 /// The predefined states when fetching data
-enum FetchingStatus { init, loading, done, error }
+enum FetchingStatus { init, loading, done, error, loadingMore }
 
 /// FetchingState allow you to handle UI base on the current state of data fetching
-class FetchingState<T, E> {
+class FetchingState<T> {
   /// fetching status [FetchingStatus]
   final FetchingStatus fetchingStatus;
 
   /// data that pass to `onDone`
   ///
-  /// useage: `FetchingState.done({T? data})`
+  /// usage: `FetchingState.done({T? data})`
   final T? data;
 
   /// data that pass to `onError`
   ///
-  /// useage: `FetchingState.error({E? error})`
-  final E? error;
+  /// usage: `FetchingState.error({Object? error})`
+  final Object? error;
 
   FetchingState({
     this.data,
@@ -26,16 +26,50 @@ class FetchingState<T, E> {
     required this.fetchingStatus,
   });
 
-  FetchingState copyWith({
+  FetchingState<T> copyWith({
     required FetchingStatus fetchingStatus,
+    T? data,
+    Object? error,
   }) {
     return FetchingState(
       fetchingStatus: fetchingStatus,
+      data: data ?? this.data,
+      error: error ?? this.error,
+    );
+  }
+
+  /// create new instance of FetchingState.
+  /// set status to [FetchingStatus.loadingMore]
+  /// keep the current [data] and [error] or user can pass in those parameter.
+  /// this is helpful when you want to add more data into current [data]
+  FetchingState<T> copyWithLoadingMore({
+    T? data,
+    Object? error,
+  }) {
+    return FetchingState(
+      fetchingStatus: FetchingStatus.loadingMore,
+      data: this.data,
+      error: this.error,
+    );
+  }
+
+  /// create new instance of FetchingState.
+  /// set status to [FetchingStatus.done]
+  /// keep the current [data] and [error] or user can pass in those parameter.
+  /// this method should be use after loading more item into data
+  FetchingState<T> copyWhenDone({
+    T? data,
+    Object? error,
+  }) {
+    return FetchingState(
+      fetchingStatus: FetchingStatus.done,
+      data: data ?? this.data,
+      error: error ?? this.error,
     );
   }
 
   /// Create new FetchingState instance, default status to [FetchingStatus.init]
-  factory FetchingState.init({T? data, E? error}) => FetchingState(
+  factory FetchingState.init({T? data, Object? error}) => FetchingState(
         fetchingStatus: FetchingStatus.init,
         data: data,
         error: error,
@@ -50,7 +84,7 @@ class FetchingState<T, E> {
       FetchingState(fetchingStatus: FetchingStatus.done, data: data);
 
   /// Create new FetchingState instance, default status to [FetchingStatus.error]
-  factory FetchingState.error([E? error]) =>
+  factory FetchingState.error([Object? error]) =>
       FetchingState(fetchingStatus: FetchingStatus.error, error: error);
 
   /// Declare methods that return widgets base on [fetchingStatus]
@@ -58,8 +92,8 @@ class FetchingState<T, E> {
   /// Allow to use [orElse], skip some methods that you don't need.
   Widget whenOrElse({
     Widget Function()? onLoading,
-    Widget Function(T? data)? onDone,
-    Widget Function(E? error)? onError,
+    Widget Function(T? data, bool isLoadingMore)? onDone,
+    Widget Function(Object? error)? onError,
     Widget Function()? onInit,
     required Widget Function()? orElse,
   }) {
@@ -68,8 +102,9 @@ class FetchingState<T, E> {
         return onInit == null ? orElse!() : onInit();
       case FetchingStatus.loading:
         return onLoading == null ? orElse!() : onLoading();
+      case FetchingStatus.loadingMore:
       case FetchingStatus.done:
-        return onDone == null ? orElse!() : onDone(data);
+        return onDone == null ? orElse!() : onDone(data, isLoadingMore);
       case FetchingStatus.error:
         return onError == null ? orElse!() : onError(error);
       default:
@@ -82,9 +117,10 @@ class FetchingState<T, E> {
   /// all state methods are required
   Widget when({
     required Widget Function() onLoading,
-    required Widget Function(T? data) onDone,
-    required Widget Function(E? error) onError,
+    required Widget Function(T? data, bool isLoadingMore) onDone,
+    required Widget Function(Object? error) onError,
     required Widget Function() onInit,
+    // required Widget Function(T? data, bool isLoadMore) onLoadMore,
   }) {
     switch (fetchingStatus) {
       case FetchingStatus.init:
@@ -92,7 +128,8 @@ class FetchingState<T, E> {
       case FetchingStatus.loading:
         return onLoading();
       case FetchingStatus.done:
-        return onDone(data);
+      case FetchingStatus.loadingMore:
+        return onDone(data, isLoadingMore);
       case FetchingStatus.error:
         return onError(error);
     }
@@ -103,8 +140,8 @@ class FetchingState<T, E> {
   /// Allow to use [orElse], skip some methods that you don't need.
   R mapOrElse<R>({
     R Function()? onLoading,
-    R Function(T? data)? onDone,
-    R Function(E? error)? onError,
+    R Function(T? data, bool isLoadingMore)? onDone,
+    R Function(Object? error)? onError,
     R Function()? onInit,
     required R Function()? orElse,
   }) {
@@ -113,8 +150,9 @@ class FetchingState<T, E> {
         return onInit == null ? orElse!() : onInit();
       case FetchingStatus.loading:
         return onLoading == null ? orElse!() : onLoading();
+      case FetchingStatus.loadingMore:
       case FetchingStatus.done:
-        return onDone == null ? orElse!() : onDone(data);
+        return onDone == null ? orElse!() : onDone(data, isLoadingMore);
       case FetchingStatus.error:
         return onError == null ? orElse!() : onError(error);
       default:
@@ -128,8 +166,8 @@ class FetchingState<T, E> {
   /// return type [R] need to be the same on all methods
   R map<R>({
     required R Function() onLoading,
-    required R Function(T? data) onDone,
-    required R Function(E? error) onError,
+    required R Function(T? data, bool isLoadingMore) onDone,
+    required R Function(Object? error) onError,
     required R Function() onInit,
   }) {
     switch (fetchingStatus) {
@@ -137,8 +175,9 @@ class FetchingState<T, E> {
         return onInit();
       case FetchingStatus.loading:
         return onLoading();
+      case FetchingStatus.loadingMore:
       case FetchingStatus.done:
-        return onDone(data);
+        return onDone(data, isLoadingMore);
       case FetchingStatus.error:
         return onError(error);
     }
@@ -152,6 +191,9 @@ class FetchingState<T, E> {
 
   /// Shortcut for checking if current status is loading
   bool get isLoading => fetchingStatus == FetchingStatus.loading;
+
+  /// Shortcut for checking if current status is loading more data
+  bool get isLoadingMore => fetchingStatus == FetchingStatus.loadingMore;
 
   /// Shortcut for checking if current status is error
   bool get isError => fetchingStatus == FetchingStatus.error;